Implemented basic arduino code and python code
It appears to be working. But I still need improve the software, basically a need to implement some sort of streaming from the analog to digital converter.

I found an error on the board: the temperature circuit was wrong. I corrected it on the physical board I built but I still need to correct the Kicad project.

Another thing that I should improve is the disposition of the digital temperature sensors. If I just switch pins I could solder a digital sensor directly on the board. May be I will doit later...