Improved the software *a lot*
Now we can choose which AI channels should be read and we can stop data acquisition in the middle of a scan!

We can also set the temperature resolution but I couldn't notice any difference in speed. I will have to study this more.

What I need to do now is improve flexibility in using the DS18b20 temperature sensors.